--- packages/rapid-cmi5/src/lib/components/modals/ScenarioSelectionModal.tsx
@@ -1,212 +1,0 @@
-/**
- * This is the generic Scenario selection modal.
- * This can be changed when using the npm package by passing in your own scenario selection modal
- **/
-
-import { useCallback } from 'react';
-import { Typography, Stack, Box, alpha, useTheme } from '@mui/material';
-import Grid from '@mui/material/Grid2';
-
-import PersonOutlineIcon from '@mui/icons-material/PersonOutline';
-import AccessTimeIcon from '@mui/icons-material/AccessTime';
-import CheckCircleIcon from '@mui/icons-material/CheckCircle';
-import ManageSearchIcon from '@mui/icons-material/ManageSearch';
-
-import { ScenarioApi } from '@rapid-cmi5/cmi5-build-common';
-import {
-  FormControlTextField,
-  DynamicModal,
-  useRangeClient,
-} from '@rapid-cmi5/ui';
-
-const ITEMS_PER_PAGE = 50;
-
-function formatDate(dateString: string): string {
-  const date = new Date(dateString);
-  const now = new Date();
-  const diffDays = Math.floor(
-    Math.abs(now.getTime() - date.getTime()) / (1000 * 60 * 60 * 24),
-  );
-
-  if (diffDays === 0) return 'Today';
-  if (diffDays === 1) return 'Yesterday';
-  if (diffDays < 7) return `${diffDays} days ago`;
-  if (diffDays < 30) return `${Math.floor(diffDays / 7)} weeks ago`;
-
-  return date.toLocaleDateString('en-US', {
-    month: 'short',
-    day: 'numeric',
-    year: date.getFullYear() !== now.getFullYear() ? 'numeric' : undefined,
-  });
-}
-
-function ScenarioCard({
-  scenario,
-  isSelected,
-}: {
-  scenario: ScenarioApi;
-  isSelected: boolean;
-}) {
-  const theme = useTheme();
-  const textPrimary = theme.palette.text.primary;
-  const textSecondary = theme.palette.text.secondary;
-  const textTertiary = alpha(theme.palette.text.secondary, 0.7);
-
-  return (
-    <Box sx={{ p: 2, pl: isSelected ? 2.5 : 2 }}>
-      <Stack spacing={1.25}>
-        <Box
-          sx={{
-            display: 'flex',
-            alignItems: 'flex-start',
-            justifyContent: 'space-between',
-            gap: 2,
-          }}
-        >
-          <Typography
-            variant="body1"
-            fontWeight={600}
-            sx={{
-              color: textPrimary,
-              lineHeight: 1.4,
-              letterSpacing: '-0.01em',
-            }}
-          >
-            {scenario.name}
-          </Typography>
-          {isSelected && (
-            <CheckCircleIcon
-              sx={{
-                fontSize: 20,
-                color: theme.palette.primary.main,
-                flexShrink: 0,
-              }}
-            />
-          )}
-        </Box>
-
-        <Stack direction="row" spacing={2} sx={{ flexWrap: 'wrap', gap: 1 }}>
-          <Stack direction="row" spacing={0.5} alignItems="center">
-            <PersonOutlineIcon sx={{ fontSize: 15, color: textTertiary }} />
-            <Typography
-              variant="caption"
-              sx={{ color: textSecondary, fontWeight: 500 }}
-            >
-              {scenario.author}
-            </Typography>
-          </Stack>
-          {scenario?.dateEdited && (
-            <Stack direction="row" spacing={0.5} alignItems="center">
-              <AccessTimeIcon sx={{ fontSize: 15, color: textTertiary }} />
-              <Typography
-                variant="caption"
-                sx={{ color: textSecondary, fontWeight: 500 }}
-              >
-                {formatDate(scenario?.dateEdited)}
-              </Typography>
-            </Stack>
-          )}
-        </Stack>
-
-        {scenario.description && (
-          <Typography
-            variant="body2"
-            sx={{
-              color: textSecondary,
-              lineHeight: 1.5,
-              display: '-webkit-box',
-              WebkitLineClamp: 2,
-              WebkitBoxOrient: 'vertical',
-              overflow: 'hidden',
-            }}
-          >
-            {scenario.description}
-          </Typography>
-        )}
-      </Stack>
-    </Box>
-  );
-}
-
-export type SubmitScenarioFormFn<T = any> = (item: T) => void;
-
-export function ScenarioSelectionForm({
-  submitForm,
-  errors,
-  control,
-}: {
-  submitForm: SubmitScenarioFormFn;
-  errors: any;
-  control: any;
-}) {
-  const rangeClient = useRangeClient();
-
-  if (!rangeClient) {
-    return (
-      <>
-        <Grid size={6}>
-          <FormControlTextField
-            control={control}
-            name={'name'}
-            required
-            label="Scenario Name"
-            error={Boolean(errors?.name)}
-            helperText={errors?.name?.message}
-          />
-        </Grid>
-        <Grid size={6}>
-          <FormControlTextField
-            control={control}
-            name={'uuid'}
-            required
-            label="Scenario UUID"
-            error={Boolean(errors?.uuid)}
-            helperText={errors?.uuid?.message}
-          />
-        </Grid>
-      </>
-    );
-  }
-  const fetchItems = useCallback(
-    async (query: { offset: number; search: string; limit: number }) => {
-      if (!rangeClient) throw Error('No client');
-
-      const response = await rangeClient.listScenarios.query({
-        query: {
-          ...query,
-          limit: ITEMS_PER_PAGE,
-          sortBy: 'dateEdited',
-          sort: 'desc',
-        },
-      });
-
-      if (response.status != 200) throw Error('Request failed');
-
-      return {
-        data: response.body.data ?? [],
-        totalCount: response.body.totalCount ?? 0,
-        totalPages: response.body.totalPages ?? 0,
-      };
-    },
-    [rangeClient],
-  );
-
-  return (
-    <DynamicModal<ScenarioApi>
-      title="Select Scenario"
-      itemLabel="scenario"
-      searchPlaceholder="Search scenarios..."
-      emptyTitle="No scenarios"
-      emptyDescription="Create your first scenario to get started"
-      itemsPerPage={ITEMS_PER_PAGE}
-      fetchItems={fetchItems}
-      getItemId={(s) => s.uuid}
-      renderItem={(s, isSelected) => (
-        <ScenarioCard scenario={s} isSelected={isSelected} />
-      )}
-      triggerLabel="Select Scenario"
-      triggerStartIcon={<ManageSearchIcon />}
-      onSelect={submitForm}
-    />
-  );
-}
