+++ packages/rapid-cmi5/src/lib/design-tools/rapidcmi5_mdx/editors/forms/scenario/ScenarioForm.tsx
@@ -28,7 +28,7 @@ import LrsHeaderWithDetails from '../LrsStatementHelper';
 
 import { toTitleCase } from '../formUtils';
 import { useEffect } from 'react';
-import { ScenarioSelectionForm } from '../../../../../components/modals/ScenarioSelectionModal';
+import { ScenarioSelectionForm } from '../../../../../components/modals/scenarios/ScenarioSelectionModal';
 import { ScenarioCard } from './ScenarioCard';
 
 export const ScenarioForm = ({
