--- packages/rapid-cmi5/src/lib/shared-modals/SharedFormModals.tsx
@@ -1,25 +1,0 @@
-
-// import {SelectionModals} from './SelectionModals';
-// import MultiSelectionModals, {} from './MultiSelectionModals';
-
-// /**
-//  * Form Modals component
-//  * Includes the common selection / view modals for this app
-//  *
-//  * @param {boolean} [isModal] Whether form is presented in a modal
-//  * @returns {React.ReactElement}
-//  */
-// /* eslint-disable react/jsx-no-useless-fragment */
-// export function SharedFormModals({ isModal = false }: { isModal?: boolean }) {
-//   return (
-//     <>
-//       {!isModal && (
-//         <>
-//           <MultiSelectionModals />
-//           <SelectionModals />
-//         </>
-//       )}
-//     </>
-//   );
-// }
-// export default SharedFormModals;
