+++ packages/ui/src/index.ts
@@ -113,3 +113,5 @@ export * from './lib/cmi5/mdx/plugins/shared/BlockAppearanceForm';
 export * from './lib/cmi5/mdx/components/AlignmentToolbarControls';
 export * from './lib/colors/BackgroundColorTrigger';
 export * from './lib/contexts/ApiContext';
+export * from './lib/utility/date';
+export * from './lib/utility/errors';
